Fixed bug where if only one gas entry the app would crash. Working on implementing the oil notification

+++ CarCare/app/src/main/java/com/apps/jlee/carcare/HomeFragment.java
@@ -221,7 +221,7 @@ public class HomeFragment extends Fragment
             public void run()
             {
                 SharedPreferences sharedpreferences = getContext().getSharedPreferences("Replacement Values", Context.MODE_PRIVATE);
-                oilPB.setProgress(sharedpreferences.getInt("oil",0));
+                oilPB.setProgress((int)sharedpreferences.getFloat("oil",0));
             }
         });
